Create NumberGen.ps1
This script randomly selects vowels and consonants to form a word of random length between 3 and 8 letters. The word generated will follow a pattern of consonant-vowel-consonant-vowel and so on. This pattern helps create words that sound like they could be real words.

Please note that this script doesn't guarantee the generation of actual English words, but it attempts to create word-like combinations based on the provided rules.

Feel free to modify the script according to your requirements or add additional rules to make the generated words more specific.